modularized which command
