+++ myshell/which.h
@@ -1,0 +1,2 @@
+void which(char *);
+int valid_which(char *);
